PNG and SVG to Icons Generator with GUI/requirements.txt
PySide6
Pillow
cairosvg
